improved basic shaping code
using floats for x1, y1, x2, y2 coordinates greatly improves shaping quality (although using `vl_font_shaper_t` is still the recommended way of rendering text)

+++ velvet/include/velvet/font/font.h
@@ -23,7 +23,7 @@ typedef struct vl_font vl_font_t;
     vl_font_new_va_expand(__VA_ARGS__, VL_SOURCE_LOCATION_HERE)
 VL_API vl_font_t *vl_font_new_(vl_platform_context_t *context, const char *name, int height, float density, const vl_byte_t *data, size_t data_length, vl_source_location_t loc);
 VL_API vl_font_atlas_codepoint_t *vl_font_rasterize_codepoint(vl_font_t *font, vl_font_atlas_t *atlas, uint32_t codepoint);
-VL_API int vl_font_get_kern_advance(vl_font_t *font, uint32_t codepoint_a, uint32_t codepoint_b);
+VL_API float vl_font_get_kern_advance(vl_font_t *font, uint32_t codepoint_a, uint32_t codepoint_b);
 VL_API vl_result_t vl_font_free(vl_font_t *font);
 
 #endif // VELVET_FONT_FONT_H
