+++ velvet/include/velvet/platform/context.h
@@ -79,7 +79,7 @@ struct vl_font_atlas_codepoint;
 struct vl_font_atlas;
 typedef struct vl_font* (*vl_ctx_font_new)(struct vl_platform_context *context, const char *name, int height, float density, const vl_byte_t *data, size_t data_length, vl_source_location_t loc);
 typedef struct vl_font_atlas_codepoint* (*vl_ctx_font_rasterize_codepoint)(struct vl_font *font, struct vl_font_atlas *atlas, uint32_t codepoint);
-typedef int (*vl_ctx_font_get_kern_advance)(struct vl_font *font, uint32_t codepoint_a, uint32_t codepoint_b);
+typedef float (*vl_ctx_font_get_kern_advance)(struct vl_font *font, uint32_t codepoint_a, uint32_t codepoint_b);
 typedef vl_result_t (*vl_ctx_font_free)(struct vl_font *font);
 
 // velvet/font/shaper.h
